+++ modules/repository-s3/src/yamlRestTest/java/org/elasticsearch/repositories/s3/RepositoryS3ClientYamlTestSuiteIT.java
@@ -9,7 +9,7 @@
 
 package org.elasticsearch.repositories.s3;
 
-import fixture.s3.Ec2ImdsHttpFixture;
+import fixture.aws.imds.Ec2ImdsHttpFixture;
 import fixture.s3.S3HttpFixture;
 import fixture.s3.S3HttpFixtureWithSessionToken;
