+++ test/fixtures/ec2-imds-fixture/src/test/java/fixture/aws/imds/Ec2ImdsHttpHandlerTests.java
@@ -7,7 +7,7 @@
  * License v3.0 only", or the "Server Side Public License, v 1".
  */
 
-package fixture.s3;
+package fixture.aws.imds;
 
 import com.sun.net.httpserver.Headers;
 import com.sun.net.httpserver.HttpContext;
